Delete synthetic.txt only if it exists
For better readability. The previous version would work too.

+++ paraphrase-job.sh
@@ -166,8 +166,12 @@ else
 
   # paraphrasing was successful, so upload the intermediate files
   aws s3 sync output_dataset s3://${s3_bucket}/${dataset_owner}/dataset/${project}/${experiment}/${output_dataset}
-  echo "Uploaded intermediate files to s3. Deleting synthetic.txt from pod..."
-  rm output_dataset/synthetic.txt
+  echo "Uploaded intermediate files to s3."
+  if test -f output_dataset/synthetic.txt ; then
+    echo "Deleting synthetic.txt from pod to save disk space..."
+    rm output_dataset/synthetic.txt
+  fi
+  
 fi
 
 if [ "$skip_filtering" = true ] ; then
